We fixed this bug with symbol "#".
If we have a tag named C#, the url is http://localhost:1313/tags/c#/ , this url show 404 page.
Symbol "#" in url is Special characters. Then we change the # to %23, then it works.

See https://github.com/Vimux/Binario/issues/28

I am not sure whether there is a better solution.

Fix #27